- Moving DB task to it's own file
