+++ tdrs-backend/tdpservice/scheduling/tasks.py
@@ -11,22 +11,10 @@ from datetime import datetime, timedelta
 import logging
 from tdpservice.email.helpers.account_access_requests import send_num_access_requests_email
 from tdpservice.email.helpers.account_deactivation_warning import send_deactivation_warning_email
-from .db_backup import run_backup
+
 
 logger = logging.getLogger(__name__)
 
-@shared_task
-def postgres_backup(*args):
-    """Run nightly postgres backup."""
-    arg = ''.join(args)
-    logger.debug("postgres_backup::run_backup() run with arg: " + arg)
-    logger.info("Begining database backup.")
-    result = run_backup(arg)
-    if result:
-        logger.info("Finished database backup.")
-    else:
-        logger.error("Failed to complete database backup.")
-    return result
 
 @shared_task
 def check_for_accounts_needing_deactivation_warning():
